Split large event_log_storage test into focused modules
