+++ backend/tests/event_log_storage.transaction.test.js
@@ -356,270 +356,4 @@ describe("event_log_storage", () => {
             expect(found).toBe(false);
         });
     });
-
-    test("getExistingEntries returns entries that were already in data.json", async () => {
-        const capabilities = getTestCapabilities();
-        await stubEventLogRepository(capabilities);
-
-        // First transaction: create initial entries
-        const firstEvent = {
-            id: { identifier: "existing1" },
-            date: new Date("2025-05-01"),
-            original: "first input",
-            input: "processed first input",
-            modifiers: { test: "first" },
-            type: "existing_event",
-            description: "First existing event",
-            creator: { name: "test", uuid: "test-uuid", version: "1.0.0" },
-        };
-
-        await transaction(capabilities, async (storage) => {
-            storage.addEntry(firstEvent, []);
-        });
-
-        // Second transaction: verify we can read existing entries and add more
-        const secondEvent = {
-            id: { identifier: "new1" },
-            date: new Date("2025-05-15"),
-            original: "new input",
-            input: "processed new input",
-            modifiers: { test: "new" },
-            type: "new_event",
-            description: "New event added after checking existing",
-            creator: { name: "test", uuid: "test-uuid", version: "1.0.0" },
-        };
-
-        await transaction(capabilities, async (storage) => {
-            // Check that we can read the existing entries
-            const existingEntries = await storage.getExistingEntries();
-            expect(existingEntries).toHaveLength(1);
-            expect(existingEntries[0].id.identifier).toEqual(
-                event.serialize(firstEvent).id
-            );
-
-            // Now add a new entry
-            storage.addEntry(secondEvent, []);
-        });
-
-        // Verify both entries are now in data.json
-        await gitstore.transaction(capabilities, async (store) => {
-            const workTree = await store.getWorkTree();
-            const dataPath = path.join(workTree, "data.json");
-            const dataFile = await capabilities.checker.instantiate(dataPath);
-            const objects = await readObjects(capabilities, dataFile);
-            expect(objects).toHaveLength(2);
-            expect(objects[0].id).toEqual(event.serialize(firstEvent).id);
-            expect(objects[1].id).toEqual(event.serialize(secondEvent).id);
-        });
-    });
-
-    test("getExistingEntries caches results to avoid repeated file reads", async () => {
-        const capabilities = getTestCapabilities();
-        await stubEventLogRepository(capabilities);
-
-        // First create an initial entry
-        const initialEvent = {
-            id: { identifier: "cache-test" },
-            date: new Date("2025-05-24"),
-            original: "cache test",
-            input: "cache test input",
-            type: "cache_test",
-            description: "Testing getExistingEntries caching",
-            creator: { name: "test", uuid: "test-uuid", version: "1.0.0" },
-        };
-
-        await transaction(capabilities, async (storage) => {
-            storage.addEntry(initialEvent, []);
-        });
-
-        // Count calls to createReadStream (already a jest mock in spies.js)
-        const reader = capabilities.reader;
-        expect(reader.createReadStream).toHaveBeenCalledTimes(0);
-
-        // Now run a new transaction to test caching
-        await transaction(capabilities, async (storage) => {
-            expect(reader.createReadStream).toHaveBeenCalledTimes(0);
-            // First call should read the file
-            const firstResult = await storage.getExistingEntries();
-            expect(firstResult).toHaveLength(1);
-            expect(reader.createReadStream).toHaveBeenCalledTimes(1);
-
-            // Second call should use the cache
-            const secondResult = await storage.getExistingEntries();
-            expect(secondResult).toHaveLength(1);
-            expect(reader.createReadStream).toHaveBeenCalledTimes(1); // Still called once.
-
-            // Both results should be identical
-            expect(secondResult).toBe(firstResult); // Same reference
-
-            // Ensure there is always something to commit
-            storage.addEntry(
-                {
-                    id: { identifier: "cache-test-2" },
-                    date: new Date("2025-05-25"),
-                    original: "cache test 2",
-                    input: "cache test input 2",
-                    type: "cache_test",
-                    description: "Testing getExistingEntries caching 2",
-                    creator: {
-                        name: "test",
-                        uuid: "test-uuid",
-                        version: "1.0.0",
-                    },
-                },
-                []
-            );
-        });
-    });
-
-    test("transaction supports config reading and writing", async () => {
-        const capabilities = getTestCapabilities();
-        await stubEventLogRepository(capabilities);
-
-        const testConfig = {
-            help: "Test configuration for transaction",
-            shortcuts: [
-                {
-                    pattern: "tx",
-                    replacement: "transaction",
-                    description: "Transaction shortcut",
-                },
-            ],
-        };
-
-        // First transaction: write config
-        await transaction(capabilities, async (storage) => {
-            storage.setConfig(testConfig);
-        });
-
-        // Second transaction: read config and verify
-        await transaction(capabilities, async (storage) => {
-            const readConfig = await storage.getExistingConfig();
-            expect(readConfig).toEqual(testConfig);
-
-            // Also verify we can read it again (caching)
-            const readConfig2 = await storage.getExistingConfig();
-            expect(readConfig2).toBe(readConfig); // Same reference due to caching
-        });
-
-        // Verify config persisted in git repository
-        await gitstore.transaction(capabilities, async (store) => {
-            const workTree = await store.getWorkTree();
-            const configPath = path.join(workTree, "config.json");
-
-            const fileExists = await fsp
-                .access(configPath)
-                .then(() => true)
-                .catch(() => false);
-            expect(fileExists).toBe(true);
-
-            const configFile = await capabilities.checker.instantiate(configPath);
-            const configStorage = require("../src/config/storage");
-            const storedConfig = await configStorage.readConfig(
-                capabilities,
-                configFile
-            );
-
-            expect(storedConfig).toEqual(testConfig);
-        });
-    });
-
-    test("transaction handles missing config.json gracefully", async () => {
-        const capabilities = getTestCapabilities();
-        await stubEventLogRepository(capabilities);
-
-        await transaction(capabilities, async (storage) => {
-            // Should not throw when config.json doesn't exist
-            const config = await storage.getExistingConfig();
-            expect(config).toBeNull();
-        });
-    });
-
-    test("transaction can update existing config", async () => {
-        const capabilities = getTestCapabilities();
-        await stubEventLogRepository(capabilities);
-
-        const initialConfig = {
-            help: "Initial config",
-            shortcuts: [{ pattern: "init", replacement: "initialize" }],
-        };
-
-        const updatedConfig = {
-            help: "Updated config",
-            shortcuts: [
-                { pattern: "init", replacement: "initialize" },
-                {
-                    pattern: "upd",
-                    replacement: "update",
-                    description: "Update shortcut",
-                },
-            ],
-        };
-
-        // Create initial config
-        await transaction(capabilities, async (storage) => {
-            storage.setConfig(initialConfig);
-        });
-
-        // Update config
-        await transaction(capabilities, async (storage) => {
-            const existing = await storage.getExistingConfig();
-            expect(existing).toEqual(initialConfig);
-
-            storage.setConfig(updatedConfig);
-        });
-
-        // Verify updated config
-        await transaction(capabilities, async (storage) => {
-            const final = await storage.getExistingConfig();
-            expect(final).toEqual(updatedConfig);
-        });
-    });
-
-    test("transaction commits when both entries and config are added", async () => {
-        const capabilities = getTestCapabilities();
-        await stubEventLogRepository(capabilities);
-
-        const testEvent = {
-            id: { identifier: "config-and-event" },
-            date: new Date("2025-05-20"),
-            original: "test with config",
-            input: "test with config",
-            type: "config_test",
-            description: "Test event with config",
-            creator: { name: "test", uuid: "test-uuid", version: "1.0.0" },
-        };
-
-        const testConfig = {
-            help: "Config with event",
-            shortcuts: [{ pattern: "evt", replacement: "event" }],
-        };
-
-        await transaction(capabilities, async (storage) => {
-            storage.addEntry(testEvent, []);
-            storage.setConfig(testConfig);
-        });
-
-        // Verify both were persisted
-        await gitstore.transaction(capabilities, async (store) => {
-            const workTree = await store.getWorkTree();
-
-            // Check data.json
-            const dataPath = path.join(workTree, "data.json");
-            const dataFile = await capabilities.checker.instantiate(dataPath);
-            const objects = await readObjects(capabilities, dataFile);
-            expect(objects).toHaveLength(1);
-            expect(objects[0].id).toBe("config-and-event");
-
-            // Check config.json
-            const configPath = path.join(workTree, "config.json");
-            const configFile = await capabilities.checker.instantiate(configPath);
-            const configStorage = require("../src/config/storage");
-            const storedConfig = await configStorage.readConfig(
-                capabilities,
-                configFile
-            );
-            expect(storedConfig).toEqual(testConfig);
-        });
-    });
 });
